move fulcio extensions into properly named folder
